pyproject.toml
[project]
name = "aegis-sentinel"
version = "1.0.0"
description = "Advanced Network EDR & IPS Gateway Sentinel Node"
authors = [{ name = "alvarofdezr" }]
requires-python = ">=3.12"
dependencies = [
    "scapy>=2.5.0",
    "netfilterqueue>=1.1.0; sys_platform == 'linux'",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]